requirements.txt
flask
flask_cors
flask_migrate
bcrypt
pyjwt
pymysql